+++ .storybook/main.js
@@ -1,17 +1,14 @@
 module.exports = {
-  "stories": [
-    "../src/**/*.stories.mdx",
-    "../src/**/*.stories.@(js|jsx|ts|tsx)"
-  ],
-  "addons": [
+  stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+  addons: [
     "@storybook/addon-links",
-    '@storybook/addon-a11y',
-    '@storybook/addon-docs',
+    "@storybook/addon-a11y",
+    "@storybook/addon-docs",
     "@storybook/addon-essentials",
     "@storybook/preset-create-react-app",
   ],
-  "framework": "@storybook/react",
-  "core": {
-    "builder": "webpack5"
-  }
-}
+  framework: "@storybook/react",
+  core: {
+    builder: "webpack5",
+  },
+};
