+++ src/Basic/Icon/index.js
@@ -1,0 +1,3 @@
+export { default as Icon } from "./Icon";
+export { default as IconBookMark } from "./IconBookMark";
+export { default as IconSocial } from "./IconSocial";
