+++ src/Basic/Logo/index.js
