+++ src/Icon/Icon.js
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
+
 import { icons } from "../shared/icons";
 
 export const Icon = ({ icon, block, ...props }) => {
