+++ src/Layouts/Navbar/Navbar.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Navbar } from "./Navbar";
+import Navbar from "./Navbar";
 
 export default {
   title: "Layouts/Navbar",
