+++ src/Layouts/Toc/Toc.jsx
@@ -10,7 +10,7 @@ const THEME = {
   LIGHT: "light",
 };
 
-export const TocNav = ({ items, ...props }) => {
+const Toc = ({ items, ...props }) => {
   const [current, setCurrent] = useState("info");
 
   return (
@@ -32,12 +32,12 @@ export const TocNav = ({ items, ...props }) => {
   );
 };
 
-TocNav.propTypes = {
+Toc.propTypes = {
   items: PropTypes.array.isRequired,
   theme: PropTypes.oneOf(Object.values(THEME)),
 };
 
-TocNav.defaultProps = {
+Toc.defaultProps = {
   theme: THEME.LIGHT,
   items: [
     {
@@ -58,10 +58,7 @@ TocNav.defaultProps = {
   ],
 };
 
-const bgColors = {
-  dark: color.black,
-  light: color.white,
-};
+export default Toc;
 
 const borderColor = {
   dark: color.gray700,
@@ -70,13 +67,6 @@ const borderColor = {
 
 const Layout = styled.div`
   display: flex;
-  align-items: flex-end;
-
   width: 100%;
-  height: 60px;
-  // TODO : 레이아웃에 맞게 맞추기
-  padding-left: 468px;
   border-bottom: 1px solid ${(props) => borderColor[props.theme]};
-
-  background-color: ${(props) => bgColors[props.theme]};
 `;
