+++ src/Layouts/Toc/index.js
@@ -1,0 +1,1 @@
+export { default as Toc } from "./Toc";
