+++ src/Layouts/TocNav/index.js
@@ -1,0 +1,1 @@
+export { default as TocNav } from "./TocNav";
